Feature/fix service

See merge request s03-final/s03p31a402!255

+++ backend/course-service/src/main/java/com/golearn/controller/TagController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,9 +17,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.golearn.dto.Tag;
+import com.golearn.service.CourseManagerService;
 import com.golearn.service.TagService;
 
 import io.swagger.annotations.ApiOperation;
+import springfox.documentation.annotations.ApiIgnore;
 
 @RestController
 @RequestMapping(value = "/tag")
@@ -26,6 +29,9 @@ public class TagController {
 
 	@Autowired
 	private TagService tagService;
+	
+	@Autowired
+	CourseManagerService courseManagerService;
 	
 	@ApiOperation(value = "코스 태그 가져오기")
 	@GetMapping(value = "/{cos_no}")
@@ -36,8 +42,11 @@ public class TagController {
 	
 	@ApiOperation(value = "코스 태그 등록하기")
 	@PostMapping(value = "/")
-	public ResponseEntity<String> registTag(@RequestBody Map<String, Object> map){
+	public ResponseEntity<String> registTag(@RequestBody Map<String, Object> map, @ApiIgnore @RequestHeader(value = "X-USERNO") String mbrNo){
 		int cosNo = (int)map.get("cos_no");
+		if(courseManagerService.checkManager(cosNo, Long.parseLong(mbrNo)) == 0) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 		List<Integer> list = (List<Integer>)map.get("list");
 		int res = tagService.insertTag(cosNo, list);
 		if(res == 0) {
@@ -48,7 +57,10 @@ public class TagController {
 	
 	@ApiOperation(value = "코스 태그 삭제")
 	@DeleteMapping(value = "/{cos_no}/{list}")
-	public ResponseEntity<String> removeTag(@PathVariable("cos_no") int cosNo, @PathVariable("list") List<Integer> list){
+	public ResponseEntity<String> removeTag(@PathVariable("cos_no") int cosNo, @PathVariable("list") List<Integer> list, @ApiIgnore @RequestHeader(value = "X-USERNO") String mbrNo){
+		if(courseManagerService.checkManager(cosNo, Long.parseLong(mbrNo)) == 0) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 		int res = tagService.deleteTag(cosNo, list);
 		if(res == 0) {
 			return ResponseEntity.ok("삭제 실패");
